// src/enums/UnitUpgrades.ts
// Units talent list, effects defined in /models/UnitTalents.ts
export enum UNIT_UPGRADES {
    // Berserk
    BERSERK_1 = "BERSERK_1", // + DAM
    BERSERK_2 = "BERSERK_2", //
    BERSERK_3 = "BERSERK_3", // + DEF
    BERSERK_4 = "BERSERK_4", // + DAM
    BERSERK_5 = "BERSERK_5", // + DAM Depending on Number

    // Bower
    BOWER_1 = "BOWER_1", // + DAM
    BOWER_2 = "BOWER_2",
    BOWER_3 = "BOWER_3", // + DEF
    BOWER_4 = "BOWER_4", // + DAM
    BOWER_5 = "BOWER_5", 

    // Guardian
    GUARDIAN_1 = "GUARDIAN_1", // + DEF
    GUARDIAN_2 = "GUARDIAN_2",
    GUARDIAN_3 = "GUARDIAN_3", // + DAM
    GUARDIAN_4 = "GUARDIAN_4", // + DEF
    GUARDIAN_5 = "GUARDIAN_5",
  }